requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
openpyxl==3.1.5
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
